Stat Manager Script
I was able to get it working with the TextMeshPro text components as opposed to the legacy text.  :)
