+++ Assets/Scripts/Employee/StatManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ef2577ec7f5b961459e9c62f5f90d9fc
+guid: 9d3870e27155a9641865f4e8fc40ccf5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
