--- Assets/Scripts/EmployeeDisplayUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db4454b893ad41f488666fd1de73e397
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
