--- Assets/Scripts/EmployeeStat.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 665db415589848a409ca8bd9f0b99f92
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
