--- Assets/Scripts/EmployeeStatModifier.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b00995b7c968e0b49a4566fbb6875a7d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
